requirements.txt
# Core dependencies
neo4j>=5.0.0,<6.0.0  # Neo4j Python driver
pydantic>=2.0.0,<3.0.0  # Data validation
python-dotenv>=1.0.0,<2.0.0  # Environment variable management

# MCP Server dependencies
mcp>=1.0.0,<2.0.0  # Model Context Protocol Python SDK with FastMCP
httpx>=0.25.0,<1.0.0  # HTTP client for async requests

# Development dependencies
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
